+++ lib/data/plugins.json
