+++ lib/views/base.njk
@@ -10,16 +10,16 @@
   <meta name="twitter:creator" content="@segment" />
   <meta name="twitter:title" content="{{title}}" />
   <meta name="twitter:description" content="{{description}}">
-  <meta name="twitter:image" content="http://www.metalsmith.io/metalsmith.io/images/twitter-image.png" />
+  <meta name="twitter:image" content="/img/twitter-image.png" />
   <meta property="og:site_name" content="Segment">
   <meta property="og:title" content="{{title}}">
   <meta property="og:description" content="{{description}}">
   <meta property="og:type" content="website">
-  <meta property="og:image" content="http://www.metalsmith.io/metalsmith.io/images/facebook-image.png" />
+  <meta property="og:image" content="/img/facebook-image.png" />
   <meta name="google-site-verification" content="2RKFuZol3mkMgto2Y0TUZUmzMpWZ-Ck-7HjocBHpJtE" />
-  <link rel="icon" href="/metalsmith.io/images/favicon.png">
+  <link rel="icon" href="/img/favicon.png">
   <link rel="stylesheet" href="//cdnjs.cloudflare.com/ajax/libs/highlight.js/9.3.0/styles/tomorrow.min.css">
-  <link rel="stylesheet" href="/build.css">
+  <link rel="stylesheet" href="/index.css">
   <link rel="stylesheet" href="//cloud.typography.com/6692672/678184/css/fonts.css">
   <link rel="stylesheet" href="//cdn.symbolset.com/94440cd0456394a198fb9c3302ad80fe007e5eff/symbolset.css">
   <meta name="node" content="{{nodeVersion}}" />
@@ -43,7 +43,7 @@
     {{ contents | safe }}
   </section>
 
-  <script type="text/javascript" src="/build.js"></script>
+  <script type="text/javascript" src="/index.js"></script>
   <script type="text/javascript">require('metalsmith.io');</script>
 </body>
 </html>
