+++ package.json
@@ -1,39 +1,16 @@
 {
   "name": "metalsmith.io",
   "private": true,
-  "engines": {
-    "node": "^6.6.0"
-  },
-  "dependencies": {
-    "blazy": "^1.8.2",
-    "component": "^0.19.9",
-    "component-builder": "~0.11.2",
-    "grunt": "^1.0.1",
-    "grunt-contrib-clean": "^1.0.0",
-    "grunt-contrib-connect": "^1.0.2",
-    "grunt-shell": "^1.3.1",
-    "metalsmith": "^1.7.0",
-    "metalsmith-in-place": "^1.3.1",
-    "metalsmith-layouts": "^1.4.1",
-    "metalsmith-markdown": "^0.2.1",
-    "metalsmith-metadata": "0.0.1",
-    "metalsmith-metallic": "^1.0.0",
-    "mkdirp": "^0.3.5",
-    "myth": "^0.1.0",
-    "swig": "^1.3.2"
+  "scripts": {
+    "build": "node metalsmith.js"
   },
   "devDependencies": {
-    "cheerio": "^0.22.0",
-    "eslint": "1.x",
-    "tape": "^4.6.0",
-    "taper": "^0.5.0"
-  },
-  "scripts": {
-    "build": "grunt build",
-    "heroku-postbuild": "npm run build",
-    "start": "grunt connect",
-    "pretest": "npm run build",
-    "test": "taper verify",
-    "posttest": "eslint --ignore-pattern node_modules --ignore-pattern build/ --ext .js '**/*.js'"
+    "jstransformer-marked": "^1.0.3",
+    "jstransformer-nunjucks": "^0.5.0",
+    "metalsmith": "^2.3.0",
+    "metalsmith-in-place": "^4.2.0",
+    "metalsmith-layouts": "^2.2.0",
+    "metalsmith-metadata": "0.0.1",
+    "metalsmith-metallic": "^1.0.0"
   }
 }
