+++ src/img/favicon.png
