+++ src/index.css
@@ -1,4 +1,3 @@
-
 /**
  * Variables.
  */
@@ -29,7 +28,7 @@
 html {
   font: 500 16px/1.6 var(sentinel);
   color: var(dark-gray);
-  background: url('images/ground.jpg') var(light-gray);
+  background: url('/img/ground.jpg') var(light-gray);
 }
 
 h1 {
@@ -118,59 +117,59 @@ hr {
 }
 
 hr:nth-of-type(1) {
-  background-image: url('images/divider-1.png');
+  background-image: url('/img/divider-1.png');
 }
 
 hr:nth-of-type(2) {
-  background-image: url('images/divider-2.png');
+  background-image: url('/img/divider-2.png');
 }
 
 hr:nth-of-type(3) {
-  background-image: url('images/divider-3.png');
+  background-image: url('/img/divider-3.png');
 }
 
 hr:nth-of-type(4) {
-  background-image: url('images/divider-4.png');
+  background-image: url('/img/divider-4.png');
 }
 
 hr:nth-of-type(5) {
-  background-image: url('images/divider-5.png');
+  background-image: url('/img/divider-5.png');
 }
 
 hr:nth-of-type(6) {
-  background-image: url('images/divider-6.png');
+  background-image: url('/img/divider-6.png');
 }
 
 hr:nth-of-type(7) {
-  background-image: url('images/divider-1.png');
+  background-image: url('/img/divider-1.png');
 }
 
 hr:nth-of-type(8) {
-  background-image: url('images/divider-2.png');
+  background-image: url('/img/divider-2.png');
 }
 
 hr:nth-of-type(9) {
-  background-image: url('images/divider-3.png');
+  background-image: url('/img/divider-3.png');
 }
 
 hr:nth-of-type(10) {
-  background-image: url('images/divider-4.png');
+  background-image: url('/img/divider-4.png');
 }
 
 hr:nth-of-type(11) {
-  background-image: url('images/divider-5.png');
+  background-image: url('/img/divider-5.png');
 }
 
 hr:nth-of-type(12) {
-  background-image: url('images/divider-6.png');
+  background-image: url('/img/divider-6.png');
 }
 
 hr:nth-of-type(13) {
-  background-image: url('images/divider-1.png');
+  background-image: url('/img/divider-1.png');
 }
 
 hr:nth-of-type(14) {
-  background-image: url('images/divider-2.png');
+  background-image: url('/img/divider-2.png');
 }
 
 /**
